chore: move preprocessing to cv/utils/imagePreprocessor, remove worklet from dispose
