+++ docs/typedoc.json
@@ -11,7 +11,6 @@
     "Modules",
     "Core / Functions",
     "Core / Types",
-    "Core / Errors",
     "Utils / Functions",
     "Utils / Types",
     "CV / Tasks",
@@ -19,6 +18,7 @@
     "CV / Constants",
     "NLP / Tasks",
     "NLP / Types",
+    "NLP / Constants",
     "Speech / Tasks",
     "Speech / Types",
     "Speech / Constants",
