+++ packages/react-native-executorch/src/extensions/cv/tasks/keypointDetection.ts
@@ -11,7 +11,7 @@ import { validateSpec, method, f32 } from '../../../core/schema';
 import { wrapAsync } from '../../../core/runtime';
 
 import type { ImageBuffer } from '../image';
-import { createImagePreprocessor, type ImagePreprocessorOptions } from './preprocessing';
+import { createImagePreprocessor, type ImagePreprocessorOptions } from '../utils/imagePreprocessor';
 
 import type { ResizeMode } from '../ops/image';
 import { scalePoint, type Point } from '../ops/points';
