+++ packages/react-native-executorch/src/extensions/cv/tasks/objectDetection.ts
@@ -12,7 +12,7 @@ import { wrapAsync } from '../../../core/runtime';
 
 import type { ResizeMode } from '../ops/image';
 import type { ImageBuffer } from '../image';
-import { createImagePreprocessor, type ImagePreprocessorOptions } from './preprocessing';
+import { createImagePreprocessor, type ImagePreprocessorOptions } from '../utils/imagePreprocessor';
 import { nms, scaleBox, decodeBox, type BoundingBox, type BoxFormat } from '../ops/boxes';
 import { RnExecuTorchError } from '../../../core/error';
