+++ packages/react-native-executorch/src/index.ts
@@ -83,7 +83,6 @@ export * from './extensions/cv/tasks/objectDetection';
 export * from './extensions/cv/tasks/imageEmbedding';
 export * from './extensions/cv/tasks/sdxsTextToImage';
 export * from './extensions/cv/tasks/paddleOcr';
-export * from './extensions/cv/tasks/preprocessing';
 export * from './extensions/llm/tasks/llmChatSession';
 export * from './extensions/nlp/tasks/tokenization';
 export * from './extensions/nlp/tasks/textEmbedding';
